+++ pep517/colorlog.py
@@ -4,6 +4,8 @@ Code copied from Tornado, Apache licensed.
 """
 # Copyright 2012 Facebook
 #
+# SPDX-License-Identifier: Apache-2.0
+#
 # Licensed under the Apache License, Version 2.0 (the "License"); you may
 # not use this file except in compliance with the License. You may obtain
 # a copy of the License at