+++ tests/samples/buildsys_pkgs/buildsys_minimal_editable.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2017-2021 Thomas Kluyver <thomas@kluyver.me.uk> and other contributors
+#
+# SPDX-License-Identifier: MIT
+
 from buildsys_minimal import build_sdist, build_wheel  # noqa
 
 build_editable = build_wheel