+++ tests/samples/pkg1/pkg1.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2017-2021 Thomas Kluyver <thomas@kluyver.me.uk> and other contributors
+#
+# SPDX-License-Identifier: MIT
+
 """Sample package for tests"""
 
 __version__ = '0.5'