+++ tests/samples/pkg_intree/backend/intree_backend.py
@@ -1,2 +1,6 @@
+# SPDX-FileCopyrightText: 2017-2021 Thomas Kluyver <thomas@kluyver.me.uk> and other contributors
+#
+# SPDX-License-Identifier: MIT
+
 def get_requires_for_build_sdist(config_settings):
     return ["intree_backend_called"]